mvp/m365_wrapper/requirements.txt
fastapi>=0.116.1
httpx>=0.27.0
python-dotenv>=1.0
uvicorn[standard]>=0.35.0
aiohttp>=3.12.0
microsoft-agents-hosting-core>=0.8.0
microsoft-agents-hosting-fastapi>=0.8.0
microsoft-agents-authentication-msal>=0.8.0
